requirements.txt
numpy
rich
pandas
joblib
auto-sklearn
tpot
autogluon.tabular 